renamend schemes.restaurant.py
